client/requirements.txt
grpcio-tools
ujson
falcon
gunicorn
tornado
tensorflow==1.0